add words comprising email to weak words list
Before this change, only the email address in its entirety is added
to the weak words list. This is a problem because many email addresses
are of the form `firstname.lastname@something` and firstname.lastname are
weak passwords.

This change adds all email address "tokens" to the weak words list.